pyproject.toml
[tool.poetry]
name = "netflix-activity"
version = "0.1.0"
description = "Download your Netflix activity"
authors = ["Nicolas Vincent"]
license = "GPLv3"
readme = "README.md"
homepage = "https://github.com/NicovincX2/netflix"
repository = "https://github.com/NicovincX2/netflix"
keywords = ["netflix"]

[tool.poetry.dependencies]
python = "^3.7"
selenium = "^3.141.0"
click = "^7.1.2"
python-dotenv = "^0.13.0"

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"
coverage = {extras = ["toml"], version = "^5.1"}
pytest-cov = "^2.10.0"
black = "^19.10b0"
flake8 = "^3.8.3"
flake8-bandit = "^2.1.2"
flake8-black = "^0.2.0"
flake8-bugbear = "^20.1.4"
flake8-import-order = "^0.18.1"

[tool.poetry.scripts]
netflix-activity = "example.main:main"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["netflix_activity"]

[tool.coverage.report]
show_missing = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
